refactor(step3): ordertable package로 이동
