+++ src/main/java/kitchenpos/ordertable/domain/JpaOrderTableRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package kitchenpos.ordertable.domain;
 
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
 }
