+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,10 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.ordertable.domain;
 
+import java.util.UUID;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.util.UUID;
 
 @Table(name = "order_table")
 @Entity
