+++ src/main/java/kitchenpos/ordertable/ui/OrderTableRestController.java
@@ -1,13 +1,18 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.OrderTableService;
-import kitchenpos.domain.OrderTable;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.ordertable.ui;
 
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.ordertable.application.OrderTableService;
+import kitchenpos.ordertable.domain.OrderTable;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/api/order-tables")
 @RestController
