+++ src/test/java/kitchenpos/Fixtures.java
@@ -12,6 +12,7 @@ import kitchenpos.domain.OrderLineItem;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 
 public class Fixtures {
